Create isometric_character_export.py
This is a script intended to be run through the Python text editor of Blender LTS version 3.6.5. It presumes that you've followed the instructions in the comments at the top of the script to setup a camera that tracks a hidden object inside of the target model while following a circular path thereabove. There must be at least one key frame for the animation for this script to work because it renders only the key frames of the animation that it finds.

Note that the scene update mid-script is not a supported feature and may be deprecated, at some point. It's considered a "hack" in the Blender documentation. See: https://docs.blender.org/api/current/info_gotcha.html