Update isometric_character_export.py
Fixed some typos in the comments. Added more explanation about the clamping requirement for the setup before running the script. Altered contribution segment about programmatically creating and position lamps. It appears that global illumination may be a built-in feature in Blender, depending on what renderer is selected and how its configured.